refactor: move push methods to pushNotification module
