+++ __tests__/wallet/wallet.test.ts
@@ -326,84 +326,6 @@ test('generateCreateWalletAuthData should return correct auth data', async () =>
   expect(authXpubMessage.verify(authXpubAddress, authData.authXpubkeySignature)).toBe(true);
 });
 
-test('registerDeviceToPushNotification', async () => {
-  const wallet = buildWalletToAuthenticateApiCall();
-  spyOn(wallet, 'isReady').and.returnValue(true);
-
-  mockAxiosAdapter.reset();
-  mockAxiosAdapter
-    .onPost('push/register')
-    .replyOnce(200, {
-      success: true,
-    })
-    .onPost('push/register')
-    .replyOnce(400, {
-      success: false,
-      error: 'invalid-payload',
-      details: [{ message: '"deviceId" length must be less than or equal to 256 characters long', path: ['deviceId'] }],
-    });
-
-  const successCall = wallet.registerDeviceToPushNotification({ deviceId: '123', pushProvider: 'android', enablePush: true });
-  
-  await expect(successCall).resolves.toStrictEqual({ success: true });
-
-  const invalidCall = wallet.registerDeviceToPushNotification({ deviceId: '123', pushProvider: 'android', enablePush: true });
-
-  await expect(invalidCall).rejects.toThrowError('Error registering device for push notification.');
-});
-
-test('updateDeviceToPushNotification', async () => {
-  const wallet = buildWalletToAuthenticateApiCall();
-  spyOn(wallet, 'isReady').and.returnValue(true);
-
-  mockAxiosAdapter.reset();
-  mockAxiosAdapter
-    .onPut('push/update')
-    .replyOnce(200, {
-      success: true,
-    })
-    .onPut('push/update')
-    .replyOnce(400, {
-      success: false,
-      error: 'invalid-payload',
-      details: [{ message: '"deviceId" length must be less than or equal to 256 characters long', path: ['deviceId'] }],
-    });
-
-  const successCall = wallet.updateDeviceToPushNotification({ deviceId: '123', enablePush: true, enableShowAmounts: true });
-  
-  await expect(successCall).resolves.toStrictEqual({ success: true });
-
-  const invalidCall = wallet.updateDeviceToPushNotification({ deviceId: '123', enablePush: true, enableShowAmounts: true });
-
-  await expect(invalidCall).rejects.toThrowError('Error updating push notification settings for device.');
-});
-
-test('unregisterDeviceToPushNotification', async () => {
-  const wallet = buildWalletToAuthenticateApiCall();
-  spyOn(wallet, 'isReady').and.returnValue(true);
-
-  mockAxiosAdapter.reset();
-  mockAxiosAdapter
-    .onPost('push/unregister')
-    .replyOnce(200, {
-      success: true,
-    })
-    .onPost('push/unregister')
-    .replyOnce(400, {
-      success: false,
-      error: 'invalid-payload',
-      details: [{ message: '"deviceId" length must be less than or equal to 256 characters long', path: ['deviceId'] }],
-    });
-
-  const successCall = wallet.unregisterDeviceToPushNotification({ deviceId: '123' });
-  
-  await expect(successCall).resolves.toStrictEqual({ success: true });
-
-  const invalidCall = wallet.unregisterDeviceToPushNotification({ deviceId: '123' });
-
-  await expect(invalidCall).rejects.toThrowError('Error unregistering wallet from push notifications.');
-});
-
 test('getTxById', async () => {
   const wallet = buildWalletToAuthenticateApiCall();
   spyOn(wallet, 'isReady').and.returnValue(true);
